Upgrade electron and electron-unhandled and remove s3 dep
New version of electron does sets nodeIntegration to be false by
default, so we have to explicitly set them to true.

Maybe regression: s3 package is old and uses graceful-fs which does not
work well with new version of electron. We should be using AWS's SDK for
speaking to s3 anyway. Given that this publisher is not live, I'm
disabling it for this upgrade